fix: restructured phone number and name-based signup api
