+++ src/apps/backend/modules/account/internal/account-reader.ts
@@ -1,7 +1,6 @@
 import {
   Account,
   AccountNotFoundError,
-  AccountWithPhoneNumberExistsError,
   AccountWithUserNameExistsError,
   InvalidCredentialsError,
   Nullable,
@@ -25,12 +24,12 @@ export default class AccountReader {
 
   public static async getAccountByUsernameAndPassword(
     password: string,
-    username: string
+    username: string,
   ): Promise<Account> {
     const account = await AccountReader.getAccountByUsername(username);
     const isPasswordValid = await AccountUtil.comparePassword(
       password,
-      account.hashedPassword
+      account.hashedPassword,
     );
     if (!isPasswordValid) {
       throw new InvalidCredentialsError(username);
@@ -52,19 +51,7 @@ export default class AccountReader {
   }
 
   public static async getAccountByPhoneNumber(
-    phoneNumber: PhoneNumber
-  ): Promise<Account> {
-    const account = await this.getAccountByPhoneNumberOptional(phoneNumber);
-
-    if (!account) {
-      throw new AccountNotFoundError(phoneNumber.toString());
-    }
-
-    return account;
-  }
-
-  public static async getAccountByPhoneNumberOptional(
-    phoneNumber: PhoneNumber
+    phoneNumber: PhoneNumber,
   ): Promise<Nullable<Account>> {
     const accountDb = await AccountRepository.findOne({
       'phoneNumber.countryCode': phoneNumber.countryCode,
@@ -80,7 +67,7 @@ export default class AccountReader {
   }
 
   public static async checkUsernameNotExists(
-    username: string
+    username: string,
   ): Promise<boolean> {
     const accountDb = await AccountRepository.findOne({
       username,
@@ -89,22 +76,6 @@ export default class AccountReader {
 
     if (accountDb) {
       throw new AccountWithUserNameExistsError(username);
-    }
-
-    return false;
-  }
-
-  public static async checkPhoneNumberNotExists(
-    phoneNumber: PhoneNumber
-  ): Promise<boolean> {
-    const accountDb = await AccountRepository.findOne({
-      'phoneNumber.countryCode': phoneNumber.countryCode,
-      'phoneNumber.phoneNumber': phoneNumber.phoneNumber,
-      active: true,
-    });
-
-    if (accountDb) {
-      throw new AccountWithPhoneNumberExistsError(phoneNumber.toString());
     }
 
     return false;
