+++ src/apps/backend/modules/account/internal/account-writer.ts
@@ -58,18 +58,16 @@ export default class AccountWriter {
   }
 
   public static async createAccountByPhoneNumberAndName(
-    firstName: string,
-    lastName: string,
     phoneNumber: PhoneNumber,
+    firstName?: string,
+    lastName?: string,
   ): Promise<Account> {
-    await this.validateAndCheckPhoneNumberExists(phoneNumber);
     const accountDb = await AccountRepository.create({
       firstName,
       lastName,
       phoneNumber,
       active: true,
     });
-
     return AccountUtil.convertAccountDBToAccount(accountDb);
   }
 
@@ -108,23 +106,5 @@ export default class AccountWriter {
 
   public static async deleteAccountById(accountId: string): Promise<void> {
     await AccountRepository.findByIdAndDelete(accountId);
-  }
-
-  private static async validateAndCheckPhoneNumberExists(
-    phoneNumber: PhoneNumber,
-  ): Promise<void> {
-    const phoneUtil = <PhoneUtilInterface>(
-      (<PhoneUtilInstance>PhoneNumberUtil).getInstance()
-    );
-    const isValidPhoneNumber = phoneUtil.isValidNumber(
-      phoneUtil.parse(phoneNumber.toString()),
-    );
-
-    if (!isValidPhoneNumber) {
-      throw new OtpRequestError('Please provide a valid phone number.');
-    }
-    // check if account already exists with the given phone number
-    // this will throw an error if it does
-    await AccountReader.checkPhoneNumberNotExists(phoneNumber);
   }
 }
