+++ src/apps/backend/modules/account/rest-api/account-controller.ts
@@ -34,11 +34,11 @@ export class AccountController {
           password,
           username
         );
-      } else if (firstName && lastName && phoneNumber) {
-        account = await AccountService.getOrCreateAccountByPhoneNumberAndName(
+      } else if ((firstName || lastName) && phoneNumber) {
+        account = await AccountService.createAccountByPhoneNumberAndName(
+          new PhoneNumber(phoneNumber.countryCode, phoneNumber.phoneNumber),
           firstName,
           lastName,
-          new PhoneNumber(phoneNumber.countryCode, phoneNumber.phoneNumber),
         );
       } else if (phoneNumber) {
         account = await AccountService.getOrCreateAccountByPhoneNumber(
